requirements.txt
# LLM 및 LangChain 관련
langchain==0.1.20
langgraph==0.0.20
langchain-core
langchain-google-genai==1.0.2

# API 요청 및 비동기 통신
httpx==0.27.0
requests==2.31.0
python-dotenv==1.0.1

# Selenium 크롤링
selenium==4.21.0

# FastAPI 서버를 추가하는 경우 (선택)
fastapi==0.110.2
uvicorn==0.29.0